backend/requirements.txt
# django
Django==4.1.7
djangorestframework==3.14.0
django-cors-headers==3.14.0 
django-imagekit==4.1.0
 
# for django Model imageField
Pillow==9.4.0 

# tests
# factory-boy==3.2.1
# Faker==17.6.0
# pytest==7.2.2
# pytest-django==4.5.2
# pytest-factoryboy==2.5.1
# coverage==7.2.1 

# formating
# black==23.1.0 
# isort==5.12.0 
# flake8==6.0.0